+++ phpunit/functional/RuleRightTest.php
@@ -71,49 +71,64 @@ class RuleRightTest extends DbTestCase
 
     public function testLocalAccount()
     {
+        $test_root_entity = $this->getTestRootEntity(true);
+
         //prepare rules
         $rules = new \RuleRight();
-        $rules_id = $rules->add([
+        $rules_id = $this->createItem(\RuleRight::class, [
             'sub_type'     => 'RuleRight',
             'name'         => 'test local account ruleright',
             'match'        => 'AND',
             'is_active'    => 1,
             'entities_id'  => 0,
             'is_recursive' => 1,
-        ]);
+        ])->getID();
 
-        $criteria = new \RuleCriteria();
-        $criteria->add([
+        $this->createItem(\RuleCriteria::class, [
             'rules_id'  => $rules_id,
             'criteria'  => 'LOGIN',
             'condition' => \Rule::PATTERN_IS,
             'pattern'   => TU_USER,
         ]);
-        $criteria->add([
+        $this->createItem(\RuleCriteria::class, [
             'rules_id'  => $rules_id,
             'criteria'  => 'MAIL_EMAIL',
             'condition' => \Rule::PATTERN_IS,
             'pattern'   => TU_USER . '@glpi.com',
         ]);
 
-        $actions = new \RuleAction();
-        $actions->add([
+        $this->createItem(\RuleAction::class, [
             'rules_id'    => $rules_id,
             'action_type' => 'assign',
             'field'       => 'profiles_id',
             'value'       => 5, // 'normal' profile
         ]);
-        $actions->add([
+        $this->createItem(\RuleAction::class, [
             'rules_id'    => $rules_id,
             'action_type' => 'assign',
             'field'       => 'entities_id',
-            'value'       => 1, // '_test_child_1' entity
+            'value'       => $test_root_entity,
         ]);
 
         // login the user to force a real synchronisation and get it's glpi id
         $this->login(TU_USER, TU_PASS, false);
         $users_id = \User::getIdByName(TU_USER);
         $this->assertGreaterThan(0, $users_id);
+
+        $user = new \User();
+        $user->getFromDB($users_id);
+        $this->assertEquals($test_root_entity, $user->fields['entities_id']);
+
+        $this->createItem(\RuleAction::class, [
+            'rules_id'    => $rules_id,
+            'action_type' => 'assign',
+            'field'       => '_entities_id_default',
+            'value'       => -1, // Full structure
+        ]);
+
+        $this->login(TU_USER, TU_PASS, false);
+        $user->getFromDB($users_id);
+        $this->assertEquals(null, $user->fields['entities_id']);
 
         // check the user got the entity/profiles assigned
         $pu = \Profile_User::getForUser($users_id, true);
@@ -122,7 +137,7 @@ class RuleRightTest extends DbTestCase
         $found = false;
         foreach ($pu as $right) {
             if (
-                isset($right['entities_id']) && $right['entities_id'] == 1
+                isset($right['entities_id']) && $right['entities_id'] == $test_root_entity
                 && isset($right['profiles_id']) && $right['profiles_id'] == 5
                 && isset($right['is_dynamic']) && $right['is_dynamic'] == 1
             ) {
@@ -153,7 +168,7 @@ class RuleRightTest extends DbTestCase
         $found = false;
         foreach ($pu as $right) {
             if (
-                isset($right['entities_id']) && $right['entities_id'] == 1
+                isset($right['entities_id']) && $right['entities_id'] == $test_root_entity
                 && isset($right['profiles_id']) && $right['profiles_id'] == 5
                 && isset($right['is_dynamic']) && $right['is_dynamic'] == 1
             ) {